requirements.txt
Django==1.8.4
argparse==1.2.1
distribute==0.6.24
django-nocaptcha-recaptcha==0.0.18
django-solo==1.1.0
requests==2.7.0
wheel==0.24.0
wsgiref==0.1.2
